Fix paragraph rendering once more.
Markdownify in hugo is a really weird beast.
Depending on that either you have one paragraph or multiple paragraphs in your markdown, it wraps elements into paragraphs or not.
Quick workaround is deleting wrapping <p> tags in places in which we have multiple paragraphs.
Nasty hack, but in that case necessary.